# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bishal's Hub API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface RegisterInput {
  email: string;
  password: string;
  name: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  user: 'user',
} as const;

export interface User {
  id: string;
  email: string;
  name: string;
  role: UserRole;
  created_at: string;
}

export interface AuthResponse {
  token: string;
  user: User;
}

export interface App {
  id: string;
  name: string;
  /** @nullable */
  url?: string | null;
  icon_url: string;
  /** @nullable */
  description?: string | null;
  created_at: string;
}

export interface AppInput {
  name: string;
  /** @nullable */
  url?: string | null;
  icon_url: string;
  /** @nullable */
  description?: string | null;
}

export interface Message {
  id: string;
  /** @nullable */
  user_id?: string | null;
  name: string;
  email: string;
  message: string;
  created_at: string;
  /** @nullable */
  reply_count?: number | null;
}

export interface MessageInput {
  name: string;
  email: string;
  message: string;
  /** @nullable */
  user_id?: string | null;
}

export type ReplySenderRole = typeof ReplySenderRole[keyof typeof ReplySenderRole];


export const ReplySenderRole = {
  admin: 'admin',
  user: 'user',
} as const;

export interface Reply {
  id: string;
  message_id: string;
  content: string;
  sender_role: ReplySenderRole;
  /** @nullable */
  sender_name?: string | null;
  created_at: string;
}

export interface ReplyInput {
  content: string;
}

export interface MessageWithReplies {
  id: string;
  /** @nullable */
  user_id?: string | null;
  name: string;
  email: string;
  message: string;
  created_at: string;
  replies: Reply[];
}

export interface AdminStats {
  total_messages: number;
  total_users: number;
  total_apps: number;
  unread_messages: number;
}

export interface DeleteResponse {
  success: boolean;
}