# lib/api-zod/src/generated/types/adminStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bishal's Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface AdminStats {
  total_messages: number;
  total_users: number;
  total_apps: number;
  unread_messages: number;
}