# lib/api-zod/src/generated/types/app.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bishal's Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface App {
  id: string;
  name: string;
  /** @nullable */
  url?: string | null;
  icon_url: string;
  /** @nullable */
  description?: string | null;
  created_at: string;
}