# lib/api-zod/src/generated/types/appInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bishal's Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface AppInput {
  name: string;
  /** @nullable */
  url?: string | null;
  icon_url: string;
  /** @nullable */
  description?: string | null;
}