# lib/api-zod/src/generated/types/messageInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bishal's Hub API
 * OpenAPI spec version: 0.1.0
 */

export interface MessageInput {
  name: string;
  email: string;
  message: string;
  /** @nullable */
  user_id?: string | null;
}