# lib/api-zod/src/generated/types/messageWithReplies.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bishal's Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { Reply } from './reply';

export interface MessageWithReplies {
  id: string;
  /** @nullable */
  user_id?: string | null;
  name: string;
  email: string;
  message: string;
  created_at: string;
  replies: Reply[];
}