# lib/api-zod/src/generated/types/reply.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Bishal's Hub API
 * OpenAPI spec version: 0.1.0
 */
import type { ReplySenderRole } from './replySenderRole';

export interface Reply {
  id: string;
  message_id: string;
  content: string;
  sender_role: ReplySenderRole;
  /** @nullable */
  sender_name?: string | null;
  created_at: string;
}